removing components from the integration example
