+++ examples/integration/demo/src/pages/index.astro
@@ -1,5 +1,4 @@
 ---
-import { MyComponent } from '@example/my-integration/components';
 ---
 
 <html lang="en">
@@ -12,6 +11,5 @@ import { MyComponent } from '@example/my-integration/components';
 	</head>
 	<body>
 		<h1>Astro</h1>
-		<MyComponent prefix="Hello, Astro!" />
 	</body>
 </html>
