--- examples/integration/packages/my-integration/components.ts
@@ -1,6 +1,0 @@
-// Do not write code directly here, instead use the `src` folder!
-// Then, use this file to export everything you want your user to access.
-
-import MyComponent from './src/MyComponent.astro';
-
-export { MyComponent };
