--- examples/integration/packages/my-integration/src/MyComponent.astro
@@ -1,8 +1,0 @@
----
-// Write your component code in this file!
-export interface Props {
-	prefix?: string;
-}
----
-
-<div>{Astro.props.prefix} My special component</div>
